+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 Changelog
 =========
 
-## Version 0.5.2 (Upcoming)
+## Version 0.5.3 (Upcoming)
+
+- new operations:
+  - `tables.pick.table`: pick a table from a `tables` instance
+  - `tables.pick.column`: pick a column from a `tables` instance
+
+## Version 0.5.2
 
 - support polars dataframe as input when creating a KiaraTable instance
 - support Jupyter preview of KiaraTable data